Package all classes into siri in src/main/java/siri
