+++ src/main/java/siri/AddDeadlineCommand.java
@@ -1,3 +1,11 @@
+package siri;
+
+import siri.Storage;
+import siri.TaskList;
+import siri.Ui;
+import siri.InvalidInputException;
+import siri.Task;
+import siri.Deadline;
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
