+++ src/main/java/siri/AddEventCommand.java
@@ -1,3 +1,5 @@
+package siri;
+
 import java.io.IOException;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
