+++ src/main/java/siri/Command.java
@@ -1,3 +1,5 @@
+package siri;
+
 import java.io.IOException;
 
 //Command.java reused from Brigette Santoso E0564307
