+++ src/main/java/siri/DeleteCommand.java
@@ -1,3 +1,5 @@
+package siri;
+
 import java.io.IOException;
 
 //DeleteCommand.java reused from Brigette Santoso E0564307
