+++ src/main/java/siri/Event.java
@@ -1,3 +1,6 @@
+package siri;
+
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
