+++ src/main/java/siri/InvalidInputException.java
@@ -1,3 +1,5 @@
+package siri;
+
 public class InvalidInputException extends Exception {
 
     public InvalidInputException(String message) {
