+++ src/main/java/siri/MarkCommand.java
@@ -1,16 +1,19 @@
+package siri;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+//MarkCommand.java reused from Brigette Santoso E0564307
+public class MarkCommand extends Command {
     protected int taskIndex;
 
-    public UnmarkCommand(int taskIndex) {
+    public MarkCommand(int taskIndex) {
         this.taskIndex = taskIndex - 1;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws InvalidInputException {
-        tasks.unmarkTask(taskIndex);
-        ui.showUnmarkTask(tasks.getTasks().get(taskIndex));
+        tasks.markTask(taskIndex);
+        ui.showMarkTask(tasks.getTasks().get(taskIndex));
         storage.save(tasks.getTasks());
     }
 }
