+++ src/main/java/siri/Parser.java
@@ -1,3 +1,5 @@
+package siri;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
