+++ src/main/java/siri/Siri.java
@@ -1,3 +1,6 @@
+package siri;
+
+import siri.*;
 import java.io.IOException;
 
 public class Siri {
