+++ src/main/java/siri/Task.java
@@ -1,3 +1,5 @@
+package siri;
+
 public class Task {
     String description;
     String initialLetter;
