+++ src/main/java/siri/Ui.java
@@ -1,3 +1,5 @@
+package siri;
+
 import java.util.ArrayList;
 import java.util.Scanner;
