+++ src/main/java/siri/UnmarkCommand.java
@@ -1,17 +1,18 @@
+package siri;
+
 import java.io.IOException;
 
-//MarkCommand.java reused from Brigette Santoso E0564307
-public class MarkCommand extends Command {
+public class UnmarkCommand extends Command {
     protected int taskIndex;
 
-    public MarkCommand(int taskIndex) {
+    public UnmarkCommand(int taskIndex) {
         this.taskIndex = taskIndex - 1;
     }
 
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) throws InvalidInputException {
-        tasks.markTask(taskIndex);
-        ui.showMarkTask(tasks.getTasks().get(taskIndex));
+        tasks.unmarkTask(taskIndex);
+        ui.showUnmarkTask(tasks.getTasks().get(taskIndex));
         storage.save(tasks.getTasks());
     }
 }
